Fix data_length values in simulationed MODE SENSE page 4 responses so that scsi_do_mode_sense() accepts them.
The data_length value does not include itself, and the pg_length value
in a mode page does not include itself or the page_code fields. We
were +1 on the former and -2 on the latter resulting in a data_length
value that was too small by one.

Eliminates 'mode sense page 5 not simulated' messages.

ok marco@