The current iommu code only touches the hardware if the bios did not enable it (I have found the code that does enable it problematic on quite a few machines, however, that's a different issue). So provide some code that so if the bios initialised the iommu for us, we'll use what it gave us. Makes iommu work on a machine of todd's.
while i'm here, we don't need to scan all pci functoins to find the
hypertransport bridge. the gart is always on function 3, so just scan
for all the bridges and not iterate over the functions too.

Thanks to todd for his infinite patience while I gave him diffs that
went ``Boom!''.