Enable IP checksum offloading in ix(4).
Note: it did not work before because the checksum offloading was
taking care about TCP and UDP but forgot about IP fragments and other
IP protocols.  We need to take care that IP fragments are handled
correctly when we do IP/TCP/UDP offloading.

ok jsg@ deraadt@, discussed with others