Attach the 2nd generation of 6000 2x2 devices even though the firmware is not yet available.
The combo WiFi/WiMAX devices need a separate firmware too (that is not
yet available), so do not use the firmware for WiFi-only devices as it
will likely not work.

Change the way active chains (the ones that have an antenna plugged in)
are detected to work around a firmware bug which could make us mark
non-existing chains as active.