expose the max number of logical volumes the hardware supports as the width of its scsibus. only advertise 1 lun on logical volumes instead of letting the midlayer fix it up to 8.
give every target on the bus max_cmds openings. iopools means they will
properly share access to them.

this in particular is useful on skinny controllers which only advertise 31
command slots. if you have 16 volumes, theyll only get 1 opening each with
the old maths. this way round the ones that are busy will share the slots.

tested on a perc5 with two volumes and hard workloads.