Change interface and span interface lists from LIST to TAILQ.
This way the configuration order is preserved.  Order matters
because MAC address are not unique: vlan interfaces can have
the same MAC as their parent.

Frames destined for the bridge itself are now delivered to
the first-configured interface that matches the MAC instead
of the last-configured.  This means that the bridge behavior
does not suddenly change anymore when adding a vlan interface.

ok henning reyk (a while ago)